WIP adding tests to importer
